# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminAuditEventResponse": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor_identifier": {
            "title": "Actor Identifier",
            "type": "string"
          },
          "actor_type": {
            "title": "Actor Type",
            "type": "string"
          },
          "after_state": {
            "additionalProperties": true,
            "title": "After State",
            "type": "object"
          },
          "before_state": {
            "additionalProperties": true,
            "title": "Before State",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "event_id": {
            "format": "uuid",
            "title": "Event Id",
            "type": "string"
          },
          "target_email": {
            "format": "email",
            "title": "Target Email",
            "type": "string"
          },
          "target_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target User Id"
          }
        },
        "required": [
          "event_id",
          "actor_type",
          "actor_identifier",
          "target_user_id",
          "target_email",
          "action",
          "before_state",
          "after_state",
          "created_at"
        ],
        "title": "AdminAuditEventResponse",
        "type": "object"
      },
      "AdminCapabilitiesResponse": {
        "properties": {
          "can_manage_quotas": {
            "title": "Can Manage Quotas",
            "type": "boolean"
          },
          "can_view_analytics": {
            "title": "Can View Analytics",
            "type": "boolean"
          },
          "can_view_operations": {
            "title": "Can View Operations",
            "type": "boolean"
          }
        },
        "required": ["can_manage_quotas", "can_view_operations", "can_view_analytics"],
        "title": "AdminCapabilitiesResponse",
        "type": "object"
      },
...
      "AdminQuotasResponse": {
        "properties": {
          "standard": {
            "$ref": "#/components/schemas/AdminQuotaResponse"
          },
          "thorough": {
            "$ref": "#/components/schemas/AdminQuotaResponse"
          }
        },
        "required": ["standard", "thorough"],
        "title": "AdminQuotasResponse",
        "type": "object"
      },
...
      "AdminUserLookupResponse": {
        "properties": {
          "quotas": {
            "$ref": "#/components/schemas/AdminQuotasResponse"
          },
          "user": {
            "$ref": "#/components/schemas/AdminUserResponse"
          }
        },
        "required": ["user", "quotas"],
        "title": "AdminUserLookupResponse",
        "type": "object"
      },
      "AdminUserResponse": {
        "properties": {
          "account_status": {
            "title": "Account Status",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": ["id", "email", "display_name", "account_status"],
        "title": "AdminUserResponse",
        "type": "object"
      },
      "BulkDeleteSearchHistoryRequest": {
        "additionalProperties": false,
        "description": "Strict owner-scoped history IDs, stably deduplicated.",
        "properties": {
          "ids": {
            "items": {
              "maximum": 9.223372036854776e18,
              "minimum": 1.0,
              "type": "integer"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Ids",
            "type": "array"
          }
        },
        "required": ["ids"],
        "title": "BulkDeleteSearchHistoryRequest",
        "type": "object"
      },
      "BulkDeleteSearchHistoryResponse": {
        "additionalProperties": false,
        "description": "Number of active owner-scoped rows soft-deleted by a bulk request.",
        "properties": {
          "deleted": {
            "minimum": 0.0,
            "title": "Deleted",
            "type": "integer"
          }
        },
        "required": ["deleted"],
        "title": "BulkDeleteSearchHistoryResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": ["message"],
        "title": "MessageResponse",
        "type": "object"
      },
      "PublicSearchFilters": {
        "additionalProperties": false,
        "description": "Filters supported by the public HTTP search contract.",
        "properties": {
          "authors": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Authors",
            "type": "array"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Categories",
            "type": "array"
          },
          "date_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date From"
          },
          "date_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date To"
          }
        },
        "title": "PublicSearchFilters",
        "type": "object"
      },
      "PublicSearchHistoryItem": {
        "additionalProperties": false,
        "description": "One public history item without internal strategy or level fields.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "elapsed_ms": {
            "minimum": 0.0,
            "title": "Elapsed Ms",
            "type": "number"
          },
          "filters": {
            "$ref": "#/components/schemas/PublicSearchFilters"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "result_count": {
            "minimum": 0.0,
            "title": "Result Count",
            "type": "integer"
          },
          "strength": {
            "$ref": "#/components/schemas/SearchStrength"
          }
        },
        "required": [
          "id",
          "query",
          "strength",
          "filters",
          "result_count",
          "elapsed_ms",
          "created_at"
        ],
        "title": "PublicSearchHistoryItem",
        "type": "object"
      },
      "PublicSearchHistoryPage": {
        "additionalProperties": false,
        "description": "A paginated search-history response with an exact snapshot total.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicSearchHistoryItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["items", "total", "limit", "offset"],
        "title": "PublicSearchHistoryPage",
        "type": "object"
      },
...
      "PublicSearchRequest": {
        "additionalProperties": false,
        "description": "Stable public search input, isolated from internal pipeline controls.",
        "examples": [
          {
            "filters": {
              "authors": [],
              "categories": ["cs.AI", "cs.IR"],
              "date_from": "2020-01-01"
            },
            "limit": 10,
            "query": "retrieval augmented generation",
            "strength": "standard"
          }
        ],
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/PublicSearchFilters"
          },
          "limit": {
            "default": 10,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "strength": {
            "$ref": "#/components/schemas/SearchStrength",
            "default": "standard"
          }
        },
        "required": ["query"],
        "title": "PublicSearchRequest",
        "type": "object"
      },
      "PublicSearchResponse": {
        "additionalProperties": false,
        "description": "Public search response; rank is authoritative and score is unnormalized.",
        "examples": [
          {
            "degraded": true,
            "elapsed_ms": 842.37,
            "hits": [
              {
                "arxiv_id": "2401.12345",
                "arxiv_url": "https://arxiv.org/abs/2401.12345",
                "authors": ["Example Author"],
                "categories": ["cs.AI", "cs.IR"],
                "pdf_url": "https://arxiv.org/pdf/2401.12345",
                "rank": 1,
                "score": 12.75,
                "submitted_at": "2024-01-20T00:00:00Z",
                "title": "A Paper About Retrieval",
                "updated_at": "2024-03-05T00:00:00Z",
                "version": 2
              }
            ],
            "query": "retrieval augmented generation",
            "result_count": 1,
            "strength": "standard"
          }
        ],
        "properties": {
          "degraded": {
            "title": "Degraded",
            "type": "boolean"
          },
          "elapsed_ms": {
            "minimum": 0.0,
            "title": "Elapsed Ms",
            "type": "number"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/PublicSearchHit"
            },
            "title": "Hits",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "result_count": {
            "minimum": 0.0,
            "title": "Result Count",
            "type": "integer"
          },
          "strength": {
            "$ref": "#/components/schemas/SearchStrength"
          }
        },
        "required": ["query", "strength", "degraded", "hits", "result_count", "elapsed_ms"],
        "title": "PublicSearchResponse",
        "type": "object"
      },
      "QuotaOverrideRequest": {
        "properties": {
          "standard": {
            "anyOf": [
              {
                "maximum": 1000000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standard"
          },
          "thorough": {
            "anyOf": [
              {
                "maximum": 1000000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thorough"
          }
        },
        "required": ["standard", "thorough"],
        "title": "QuotaOverrideRequest",
        "type": "object"
      },
      "QuotaOverrideUpdateResponse": {
        "properties": {
          "changed": {
            "title": "Changed",
            "type": "boolean"
          }
        },
        "required": ["changed"],
        "title": "QuotaOverrideUpdateResponse",
        "type": "object"
      },
...
      "SearchStrength": {
        "description": "Public search quality presets.",
        "enum": ["standard", "thorough"],
        "title": "SearchStrength",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-powered academic paper search engine",
    "title": "Scholight API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/audit-events": {
      "get": {
        "operationId": "audit_events_admin_audit_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminAuditEventResponse"
                  },
                  "title": "Response Audit Events Admin Audit Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Audit Events",
        "tags": ["admin"]
      }
    },
    "/admin/capabilities": {
      "get": {
        "operationId": "capabilities_admin_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCapabilitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Capabilities",
        "tags": ["admin"]
      }
    },
...
    "/admin/users/lookup": {
      "get": {
        "operationId": "lookup_user_admin_users_lookup_get",
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "required": true,
            "schema": {
              "format": "email",
              "title": "Email",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserLookupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Lookup User",
        "tags": ["admin"]
      }
    },
    "/admin/users/{user_id}/quota-overrides": {
      "put": {
        "operationId": "update_quota_overrides_admin_users__user_id__quota_overrides_put",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuotaOverrideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaOverrideUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Update Quota Overrides",
        "tags": ["admin"]
      }
    },
...
    "/search": {
      "post": {
        "operationId": "search_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {}
        ],
        "summary": "Search",
        "tags": ["search"]
      }
    },
    "/search/history": {
      "get": {
        "operationId": "list_history_search_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicSearchHistoryPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List History",
        "tags": ["search"]
      }
    },
    "/search/history/bulk-delete": {
      "post": {
        "operationId": "bulk_delete_history_search_history_bulk_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteSearchHistoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteSearchHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Bulk Delete History",
        "tags": ["search"]
      }
    },
    "/search/history/{entry_id}": {
      "delete": {
        "operationId": "delete_history_entry_search_history__entry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete History Entry",
        "tags": ["search"]
      }
    },
...
  }
}